Add initial files for all folders
