+++ lib/data/app_descriptions.dart
@@ -1,0 +1,1 @@
+const String title = 'Wordiest';
