+++ lib/models/point.dart
@@ -1,0 +1,1 @@
+class Point {}
