+++ lib/models/word.dart
@@ -1,0 +1,1 @@
+class Word {}
