+++ lib/utilities/score.dart
@@ -1,4 +1,4 @@
-class Counter {
+class Score {
   int value = 0;
 
   void increment() => value++;
