+++ test/widget_tests/home_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:wordiest/main.dart';
+import 'package:wordiest/screens/home.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const Home(title: ''));
 
     expect(find.text('0'), findsOneWidget);
     expect(find.text('1'), findsNothing);
